+++ aikau/src/main/resources/alfresco/core/topics.js
@@ -924,10 +924,6 @@ define([],function() {
        * @since 1.0.71
        *
        * @event
-       * @property {Function} [displayCallback] This function will be called  once the indicator is on-screen. This can
-       *                                        be used to ensure that the indicator is fully displayed before any
-       *                                        intensive JS could cause the display transition to halt, potentially
-       *                                        preventing the indicator from displaying.
        */
       PROGRESS_INDICATOR_ADD_ACTIVITY: "ALF_PROGRESS_INDICATOR_ADD_ACTIVITY",