+++ aikau/src/test/resources/testApp/WEB-INF/classes/alfresco/site-webscripts/alfresco/services/NavigationService.get.js
@@ -34,7 +34,12 @@ model.jsonModel = {
             }
       },
       "alfresco/services/NavigationService",
-      "alfresco/services/NotificationService"
+      {
+         name: "alfresco/services/NotificationService",
+         config: {
+            showProgressIndicator: true
+         }
+      }
    ],
    widgets: [
       label,
@@ -248,7 +253,27 @@ model.jsonModel = {
                         
                      }
                   }
+               },
+               {
+                  id: "SAME_PAGE_BUT_WITH_HASHES",
+                  name: "alfresco/menus/AlfMenuBarItem",
+                  config: {
+                     label: "Page relative no site with hash",
+                     publishTopic: "ALF_NAVIGATE_TO_PAGE",
+                     publishPayload: {
+                        url: "tp/ws/NavigationService#hash=test",
+                        type: "PAGE_RELATIVE",
+                        target: "CURRENT"
+                     },
+                     publishPayloadType: "PROCESS",
+                     publishPayloadModifiers: ["processCurrentItemTokens"],
+                     publishPayloadItemMixin: true,
+                     currentItem: {
+                        
+                     }
+                  }
                }
+
             ]
          }
       },