+++ aikau/src/test/resources/testApp/WEB-INF/classes/alfresco/site-webscripts/alfresco/services/actions/ManageAspects.get.js
@@ -9,7 +9,6 @@ model.jsonModel = {
                }
             }
       },
-      "alfresco/services/NotificationService",
       "alfresco/services/DialogService",
       "alfresco/services/ActionService",
       {